Created api mocks package and refactored code to use package
